moved test location
