+++ UHDD/src/test/application/CredentialManagerTest.java
@@ -1,3 +1,4 @@
+package application;
 /*
 package application.test.application;
