+++ UHDD/src/test/application/PasswordHashTest.java
@@ -1,3 +1,4 @@
+package application;
 /*
 package application.test.application;
